+++ src/components/data-exploration/code-based/CodeBasedExploration.tsx
@@ -64,16 +64,15 @@ const CodeBasedExplorationPage: React.FC<{
                         <Typography>
                             These datasets are most easily accessed via python
                             notebooks hosted on Google's Colaboratory, for which
-                            we have several examples listed below. For approved
-                            users, CIDC will allow data access and pay for all
-                            costs associated with running these analyses. Please
+                            we have several examples listed below. CIDC will
+                            allow data access for all approved users. Please
                             contact CIDC at{" "}
                             <a href="mailto:cidc@jimmy.harvard.edu">
                                 cidc@jimmy.harvard.edu
                             </a>{" "}
-                            for access to the data tables and analysis hosting.
-                            The same data restrictions and security apply to all
-                            data accessed via this mechanism as via the portal.
+                            for access to the data tables. The same data
+                            restrictions and security apply to all data accessed
+                            via this mechanism as via the portal.
                         </Typography>
                     </CardContent>
                 </Card>